requirements.txt
tensorflow-gpu
pandas
numpy
psycopg2-binary
meteostat
scikit-learn
pyyaml
h5py
